requirements.txt
openai
elasticsearch
pdfplumber
requests
gradio
python-dotenv
PyPDF2
transformers